requirements.txt
psycopg2
sqlalchemy
pydantic
fastapi
uvicorn[standard]
requests
python-jose
cachetools
